Reimplement insert/remove as recursive; add recompression
Recompression requires keeping track of parents as
you descend the tree, so they can be compressed in
reverse order after insertion/removal. This can be
done with some kind of stack - for example, a heap-
allocated vector, or the call stack itself.

In this case I've chosen to use the call stack,
implementing the functions recursively. Different
stacking mechanisms like vectors could be explored in
the future when we have some real-world benchmarks.